+++ cpp-test/oracle/oracle.h
@@ -139,5 +139,6 @@ int run_fmpz_poly_evaluate_horner_fmpq(const char * arg);
 int run_fmpz_poly_evaluate_divconquer_fmpq(const char * arg);
 int run_fmpq_poly_evaluate_fmpq(const char * arg);
 int run_fmpq_poly_evaluate_fmpz(const char * arg);
+int run_fmpz_mod_poly_evaluate_fmpz(const char * arg);
 
 #endif