+++ malachite-nz/src/bin_util/demo_and_bench/natural_polynomial/arithmetic/evaluate.rs
@@ -6,18 +6,21 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-use malachite_base::polynomial::Evaluate;
+use malachite_base::polynomial::{Evaluate, EvaluateModPowerOf2};
 use malachite_base::test_util::bench::{BenchmarkType, run_benchmark};
 use malachite_base::test_util::generators::common::{GenConfig, GenMode};
 use malachite_base::test_util::runner::Runner;
 use malachite_nz::integer_polynomial::arithmetic::evaluate::{
     evaluate_divide_and_conquer, evaluate_horner,
 };
-use malachite_nz::test_util::bench::bucketers::pair_1_natural_polynomial_bit_bucketer;
+use malachite_nz::test_util::bench::bucketers::{
+    pair_1_natural_polynomial_bit_bucketer, triple_1_natural_polynomial_bit_bucketer,
+};
 use malachite_nz::test_util::generators::{
     natural_polynomial_natural_pair_gen, natural_polynomial_natural_pair_gen_var_2,
+    natural_polynomial_natural_unsigned_triple_gen_var_1,
 };
-use malachite_nz::test_util::natural_polynomial::arithmetic::evaluate::evaluate_naive;
+use malachite_nz::test_util::natural_polynomial::arithmetic::evaluate::*;
 
 pub(crate) fn register(runner: &mut Runner) {
     register_demo!(runner, demo_natural_polynomial_evaluate);
@@ -38,6 +41,16 @@ pub(crate) fn register(runner: &mut Runner) {
     register_bench!(
         runner,
         benchmark_natural_polynomial_evaluate_algorithms_long
+    );
+    register_demo!(runner, demo_natural_polynomial_evaluate_mod_power_of_2);
+    register_demo!(runner, demo_natural_polynomial_evaluate_mod_power_of_2_ref);
+    register_bench!(
+        runner,
+        benchmark_natural_polynomial_evaluate_mod_power_of_2_evaluation_strategy
+    );
+    register_bench!(
+        runner,
+        benchmark_natural_polynomial_evaluate_mod_power_of_2_algorithms
     );
 }
 
@@ -190,6 +203,95 @@ fn benchmark_natural_polynomial_evaluate_algorithms_long(
                 no_out!(evaluate_divide_and_conquer(p.coefficients_asc(), &x));
             }),
             ("naive", &mut |(p, x)| no_out!(evaluate_naive(&p, &x))),
+        ],
+    );
+}
+
+fn demo_natural_polynomial_evaluate_mod_power_of_2(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (p, x, pow) in natural_polynomial_natural_unsigned_triple_gen_var_1()
+        .get(gm, config)
+        .take(limit)
+    {
+        let x_old = x.clone();
+        println!(
+            "(&({p})).evaluate_mod_power_of_2({x_old}, {pow}) = {}",
+            (&p).evaluate_mod_power_of_2(x, pow)
+        );
+    }
+}
+
+fn demo_natural_polynomial_evaluate_mod_power_of_2_ref(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+) {
+    for (p, x, pow) in natural_polynomial_natural_unsigned_triple_gen_var_1()
+        .get(gm, config)
+        .take(limit)
+    {
+        println!(
+            "(&({p})).evaluate_mod_power_of_2({x}, {pow}) = {}",
+            (&p).evaluate_mod_power_of_2(&x, pow)
+        );
+    }
+}
+
+fn benchmark_natural_polynomial_evaluate_mod_power_of_2_evaluation_strategy(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+    file_name: &str,
+) {
+    run_benchmark(
+        "(&NaturalPolynomial).evaluate_mod_power_of_2(Natural, u64)",
+        BenchmarkType::EvaluationStrategy,
+        natural_polynomial_natural_unsigned_triple_gen_var_1().get(gm, config),
+        gm.name(),
+        limit,
+        file_name,
+        &triple_1_natural_polynomial_bit_bucketer("p"),
+        &mut [
+            (
+                "(&NaturalPolynomial).evaluate_mod_power_of_2(Natural, u64)",
+                &mut |(p, x, pow)| no_out!((&p).evaluate_mod_power_of_2(x, pow)),
+            ),
+            (
+                "(&NaturalPolynomial).evaluate_mod_power_of_2(&Natural, u64)",
+                &mut |(p, x, pow)| no_out!((&p).evaluate_mod_power_of_2(&x, pow)),
+            ),
+            (
+                "NaturalPolynomial.evaluate_mod_power_of_2(Natural, u64)",
+                &mut |(p, x, pow)| no_out!(p.evaluate_mod_power_of_2(x, pow)),
+            ),
+            (
+                "NaturalPolynomial.evaluate_mod_power_of_2(&Natural, u64)",
+                &mut |(p, x, pow)| no_out!(p.evaluate_mod_power_of_2(&x, pow)),
+            ),
+        ],
+    );
+}
+
+fn benchmark_natural_polynomial_evaluate_mod_power_of_2_algorithms(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+    file_name: &str,
+) {
+    run_benchmark(
+        "(&NaturalPolynomial).evaluate_mod_power_of_2(&Natural, u64)",
+        BenchmarkType::Algorithms,
+        natural_polynomial_natural_unsigned_triple_gen_var_1().get(gm, config),
+        gm.name(),
+        limit,
+        file_name,
+        &triple_1_natural_polynomial_bit_bucketer("p"),
+        &mut [
+            ("default", &mut |(p, x, pow)| {
+                no_out!((&p).evaluate_mod_power_of_2(&x, pow));
+            }),
+            ("evaluate, then reduce", &mut |(p, x, pow)| {
+                no_out!(evaluate_mod_power_of_2_naive(&p, &x, pow));
+            }),
         ],
     );
 }